added some user service
